Former-commit-id: 6552e9e0a690631b941a9d3cdb72f0f21b2bfd67 [formerly 60ae505a6ea2b94bf300c3b74ba85c6988d8c646]
Former-commit-id: b1e49e497a4b67b391fa7e9ce0ad2be15a4258d9

+++ backend/RunWithMe/src/main/java/kr/co/rwm/service/ReplyServiceImpl.java
@@ -1,6 +1,7 @@
 package kr.co.rwm.service;
 
 import java.util.Date;
+
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -10,64 +11,95 @@ import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import kr.co.rwm.entity.Board;
 import kr.co.rwm.entity.Reply;
 import kr.co.rwm.entity.User;
+import kr.co.rwm.repo.BoardRepository;
+import kr.co.rwm.repo.RanksRepository;
+import kr.co.rwm.repo.RecordRepository;
 import kr.co.rwm.repo.ReplyRepository;
+import kr.co.rwm.repo.RunningRepository;
+import kr.co.rwm.repo.RunningUserRepository;
 import kr.co.rwm.repo.UserRepository;
+import lombok.RequiredArgsConstructor;
 
+@RequiredArgsConstructor
 @Service
-public class ReplyServiceImpl implements ReplyService{
+public class ReplyServiceImpl implements ReplyService {
 
+	private final BoardRepository boardRepository;
+	
 	@Autowired
 	ReplyRepository replyRepository;
-	
+
 	@Autowired
 	UserRepository userRepository;
-	
+
 	@Override
 	public List<Reply> allReplyList() {
-		
+
 		List<Reply> list = replyRepository.findAll();
 		return list;
 	}
 
 	@Override
 	public Reply save(int uid, Map<String, String> replyInfo) {
-		
-		Reply reply = Reply.builder()
-				.boardId(Integer.parseInt(replyInfo.get("boardId")))
-				.content(replyInfo.get("content"))
-				.replyDepth(Integer.parseInt(replyInfo.get("replyDepth")))
+
+		Reply reply = Reply.builder().boardId(Integer.parseInt(replyInfo.get("boardId")))
+				.content(replyInfo.get("content")).replyDepth(Integer.parseInt(replyInfo.get("replyDepth")))
 				.parentId(Integer.parseInt(replyInfo.get("parentId")))
-				.replyOrder(Integer.parseInt(replyInfo.get("replyOrder")))
-				.regdate(new Date())
-				.editdate(new Date())
+				.replyOrder(Integer.parseInt(replyInfo.get("replyOrder"))).regdate(new Date()).editdate(new Date())
 				.build();
-		
-		Optional<User> user = userRepository.findByUserId(uid);
-		reply.setUser(user.get());
-		replyRepository.save(reply);
-		return reply;
+
+		Optional<Board> board = boardRepository.findByBoardId(Integer.parseInt(replyInfo.get("boardId")));
+		if(board.isPresent()) {
+			board.get().setReplyCount(board.get().getReplyCount()+1);
+			
+			Optional<User> user = userRepository.findByUserId(uid);
+			if(user.isPresent()) {
+				reply.setUser(user.get());
+				replyRepository.save(reply);
+				return reply;
+			}else {
+				return null;
+			}
+		}else {
+			return null;
+		}
 	}
 
 	@Override
 	@Transactional
 	public Reply update(Map<String, String> replyInfo) {
 		Optional<Reply> reply = replyRepository.findByReplyId(Integer.parseInt(replyInfo.get("replyId")));
-		 
-		if(!reply.isPresent()) {
+
+		if (!reply.isPresent()) {
 			return null;
+		}else {
+			reply.get().setContent(replyInfo.get("content"));
+			reply.get().setEditdate(new Date());
+			return reply.get();
 		}
-		reply.get().setContent(replyInfo.get("content"));
-		reply.get().setEditdate(new Date());
-		return reply.get();
 	}
 
 	@Override
 	@Transactional
 	public Long delete(int reply_id) {
-		Long ret = replyRepository.deleteByReplyId(reply_id);
-		return ret;
+		int boardId = replyRepository.findByReplyId(reply_id).get().getBoardId();
+		Optional<Board> board = boardRepository.findByBoardId(boardId);
+		if(board.isPresent()) {
+			board.get().setReplyCount(board.get().getReplyCount()-1);
+			Long ret = replyRepository.deleteByReplyId(reply_id);
+			return ret;
+			
+		}else {
+			return null;
+		}
 	}
-	
+
+	@Override
+	public List<Reply> findReplyByBoardId(int boardId) {
+		List<Reply> reply = replyRepository.findAllByBoardId(boardId);
+		return reply;
+	}
 }
